Migrate what is left from snapshot testing to check-dts
