+++ check-dts/converter.errors.ts
@@ -1,0 +1,8 @@
+import converter from '../src/converter';
+import type { Rgb } from '../src/rgb/types';
+
+// THROWS Type 'Rgb | undefined' is not assignable to type 'Rgb'.
+const case_1_expect_error: Rgb = converter()('#f0f0f0');
+
+// THROWS Type 'Lab | undefined' is not assignable to type 'Rgb'.
+const case_3_expect_error: Rgb = converter('lab')('#f0f0f0');
