--- test-d-ts/test-tsconfig.json
@@ -1,17 +1,0 @@
-{
-	"include": ["./**/*.ts"],
-	"compilerOptions": {
-		"target": "es2020",
-		"module": "commonjs",
-		"rootDir": ".",
-		"allowJs": false,
-		"declaration": true,
-		"importsNotUsedAsValues": "error",
-		"noEmit": true,
-		"isolatedModules": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"strict": true,
-		"skipLibCheck": false
-	}
-}
